Fix GHSA-v95c-p5hm-xq8f
go mod why and go mod list all doesn't really explain why this package
is defined in go.mod, but it's there after a go mod tidy.

The reason might be because the go compiler optimizes that dependency
away.

Given that the dependency has a known vulnerability, I've bumped
the version to the patched version regardless.