bug(nadatokens)
We periodically sync team names from nais console and they have removed the
nais-team prefix from the slug name returned. On our end this has caused us to create multiple nada tokens for some teams created earlier.
In addition the wrong team token is the one that has been visible for the user through the ui when logged in to markedsplassen.

The migration in this commit will ensure that each team only has one nada-token, and that this is the one that is shown in the ui
for the user.