chore(amplitude event publish timeout)
We have occationally experienced that publishing events for story
views to amplitude is unstable due to amplitude-proxy. This causes
nada-backend to wait in worst case up to 30 seconds before returning the
story files when it is accessed through markedsplassen.

Add 1 second timeout to amplitude event publish for quarto story view