bug(wait for database creation)
When we add a new bigquery datasource in metabase it is by default
granted all users access. Our metabase sync restricts access to a
different permission group when the database added is not
intended to be accessible for all users.

However, occationally we see that the check for whether the
database has been created times out and the metabase sync exits before
restricing access which then unintentionally grants all users access to a dataset
intended to be restricted.

This commit will double  the wait timeout and also
cleanup by removing database, third party mapping in nada backend db and
also permission group, collection and gcp service accounts if waiting
for database created times out.