feat(metabasesync): sync table visibility for metabase databases
Previously we have hidden all other tables in a bigquery dataset when
adding a markedsplassen dataset to metabase. When the table is
accessible by all users there exists a use case where users might want
to join with other tables or views in the same bigquery dataset.

This commit alters the sync we have had for hiding other tables in the
same bigquery dataset to instead make all all-users accessible tables or
views in the same bigquery dataset queryable in a metabase database.

Restricted datasets are handled the same way as earlier, i.e. all
other tables/views are hidden.